Se refactoriza el código en archivos por funcionalidad para mejorar la legibilidad y la modularidad
